fix shaders, use float instead of double

+++ tasks/local_shadertoy2/App.cpp
@@ -251,7 +251,7 @@ void App::drawFrame()
       Constants constants {
         .res = res,
         .cursor = cursor,
-        .time = time,
+        .time = (float)time,
       };
